LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-activation-specified-direction-position-expected.txt

Introduction

Lorem ipsum

Lorem ipsum

Focus Behavior

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Roles

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum

Lorem ipsum


FAIL CSS Overflow Test: ::scroll-marker activation scrolls into view with doesn't override specified direction position with `start` assert_equals: scroll marker activation shouldn scroll into view with specified direction position `end` and not default `start` expected 277 but got 0
